requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.0.9
xlrd>=2.0.1

# Web framework
flask>=2.2.0
flask-cors>=3.0.10

# Data processing
pyyaml>=6.0
python-dateutil>=2.8.2

# Logging and configuration
colorlog>=6.6.0
configparser>=5.3.0

# Performance and caching
joblib>=1.2.0
diskcache>=5.4.0

# Visualization support (for future enhancements)
plotly>=5.11.0
matplotlib>=3.6.0

# Development dependencies (optional)
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.10.0
flake8>=5.0.4
mypy>=0.991

# Documentation (optional)
sphinx>=5.3.0
sphinx-rtd-theme>=1.1.1
